--- a/neo/renderer/NVRHI/RenderBackend_NVRHI.cpp
+++ b/neo/renderer/NVRHI/RenderBackend_NVRHI.cpp
@@ -448,7 +448,7 @@ void idRenderBackend::DrawElementsWithCounters( const drawSurf_t* surf, bool sha
 
 		for( int i = 0; i < layouts->Num(); i++ )
 		{
-			if( !currentBindingSets[i] || *currentBindingSets[i]->getDesc() != pendingBindingSetDescs[bindingLayoutType][i] )
+			if( !currentBindingSets[i] || *currentBindingSets[i]->getDesc() != pendingBindingSetDescs[bindingLayoutType][i] || bindingLayoutType != prevBindingLayoutType )
 			{
 				currentBindingSets[i] = bindingCache.GetOrCreateBindingSet( pendingBindingSetDescs[bindingLayoutType][i], ( *layouts )[i] );
 				changeState = true;
@@ -482,9 +482,17 @@ void idRenderBackend::DrawElementsWithCounters( const drawSurf_t* surf, bool sha
 	{
 		if( deviceManager->GetGraphicsAPI() == nvrhi::GraphicsAPI::VULKAN )
 		{
-			// Reset the graphics state if the constant buffer is written to since
-			// the render pass is ended for vulkan. setGraphicsState will
-			// reinstate the render pass.
+			// Reset the graphics state if push constants are enabled or
+			// the constant buffer is written to and the render pass is
+			// ended for vulkan. setGraphicsState will reinstate the
+			// render pass or set up for push constants.
+			changeState = true;
+		}
+		else if( renderProgManager.layoutAttributes[bindingLayoutType].pcEnabled )
+		{
+			// Reset the graphics state if push constants are enabled and
+			// the uniforms or binding layout have changed for DX12.
+			// setGraphicsState will set up for push constants.
 			changeState = true;
 		}
 	}
@@ -524,6 +532,8 @@ void idRenderBackend::DrawElementsWithCounters( const drawSurf_t* surf, bool sha
 		}
 
 		commandList->setGraphicsState( state );
+
+		renderProgManager.CommitPushConstants( commandList, bindingLayoutType );
 	}
 
 	//
@@ -564,19 +574,24 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 	nvrhi::IBuffer* paramCb = renderProgManager.ConstantBuffer();
 	auto range = nvrhi::EntireBuffer;
 
-	if( type == BINDING_LAYOUT_DEFAULT )
+	auto uniformsBindingSetItem = nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range );
+	if( deviceManager->m_DeviceParams.maxPushConstantSize >= renderProgManager.layoutAttributes[type].rpBufSize )
+	{
+		uniformsBindingSetItem = nvrhi::BindingSetItem::PushConstants( 0, renderProgManager.layoutAttributes[type].rpBufSize );
+	}
+
+	if( type == BINDING_LAYOUT_DEFAULT || type == BINDING_LAYOUT_GBUFFER || type == BINDING_LAYOUT_TEXTURE || type == BINDING_LAYOUT_WOBBLESKY || type == BINDING_LAYOUT_SSGI || type == BINDING_LAYOUT_POST_PROCESS )
 	{
 		if( desc[0].bindings.empty() )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 
 		if( desc[1].bindings.empty() )
@@ -603,21 +618,20 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 			desc[2].bindings[0].resourceHandle = ( nvrhi::ISampler* )GetImageAt( 0 )->GetSampler( samplerCache );
 		}
 	}
-	else if( type == BINDING_LAYOUT_DEFAULT_SKINNED )
+	else if( type == BINDING_LAYOUT_DEFAULT_SKINNED || type == BINDING_LAYOUT_GBUFFER_SKINNED || type == BINDING_LAYOUT_TEXTURE_SKINNED || type == BINDING_LAYOUT_SSGI_SKINNED )
 	{
 		if( desc[0].bindings.empty() )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -652,13 +666,12 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 	}
 	else if( type == BINDING_LAYOUT_CONSTANT_BUFFER_ONLY_SKINNED )
@@ -667,15 +680,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -686,13 +698,12 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 
 		if( desc[1].bindings.empty() )
@@ -753,15 +764,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -824,7 +834,7 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::Texture_SRV( 0, ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 1, ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 2, ( nvrhi::ITexture* )GetImageAt( 2 )->GetTextureID() )
@@ -832,8 +842,7 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 			desc[0].bindings[1].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID();
 			desc[0].bindings[2].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID();
 			desc[0].bindings[3].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 2 )->GetTextureID();
@@ -858,13 +867,12 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 
 		// materials: 1
@@ -921,15 +929,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -988,13 +995,12 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 
 		// materials: 1
@@ -1064,15 +1070,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -1143,13 +1148,12 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 
 		if( desc[1].bindings.empty() )
@@ -1188,15 +1192,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -1237,13 +1240,12 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 
 		if( desc[1].bindings.empty() )
@@ -1279,15 +1281,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -1325,7 +1326,7 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::Texture_SRV( 0, ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 1, ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 2, ( nvrhi::ITexture* )GetImageAt( 2 )->GetTextureID() )
@@ -1333,8 +1334,7 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 			desc[0].bindings[1].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID();
 			desc[0].bindings[2].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID();
 			desc[0].bindings[3].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 2 )->GetTextureID();
@@ -1358,15 +1358,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::Texture_SRV( 0, ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 1, ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID() )
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 			desc[0].bindings[1].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID();
 			desc[0].bindings[2].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID();
 		}
@@ -1391,13 +1390,12 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 		}
 
 		if( desc[1].bindings.empty() )
@@ -1433,16 +1431,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0,  paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::StructuredBuffer_SRV( 11, currentJointBuffer, nvrhi::Format::UNKNOWN, nvrhi::BufferRange( currentJointOffset, sizeof( idVec4 ) * numBoneMatrices ) )
 			};
 		}
 		else
 		{
 			auto& bindings = desc[0].bindings;
-			bindings[0].resourceHandle = paramCb;
-			bindings[0].range = range;
-
+			bindings[0] = uniformsBindingSetItem;
 			bindings[1].resourceHandle = currentJointBuffer;
 			bindings[1].range = nvrhi::BufferRange{ currentJointOffset, sizeof( idVec4 )* numBoneMatrices };
 		}
@@ -1480,7 +1476,7 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::Texture_SRV( 0, ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 1, ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 2, ( nvrhi::ITexture* )GetImageAt( 2 )->GetTextureID() )
@@ -1488,8 +1484,7 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 			desc[0].bindings[1].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID();
 			desc[0].bindings[2].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID();
 			desc[0].bindings[3].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 2 )->GetTextureID();
@@ -1513,15 +1508,14 @@ void idRenderBackend::GetCurrentBindingLayout( int type )
 		{
 			desc[0].bindings =
 			{
-				nvrhi::BindingSetItem::ConstantBuffer( 0, paramCb, range ),
+				uniformsBindingSetItem,
 				nvrhi::BindingSetItem::Texture_SRV( 0, ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID() ),
 				nvrhi::BindingSetItem::Texture_SRV( 1, ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID() )
 			};
 		}
 		else
 		{
-			desc[0].bindings[0].resourceHandle = paramCb;
-			desc[0].bindings[0].range = range;
+			desc[0].bindings[0] = uniformsBindingSetItem;
 			desc[0].bindings[1].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 0 )->GetTextureID();
 			desc[0].bindings[2].resourceHandle = ( nvrhi::ITexture* )GetImageAt( 1 )->GetTextureID();
 		}