This PR replaces the performance part of https://github.com/RobertBeckebans/RBDOOM-3-BFG/pull/818, which will be closed and not merged. It has one dependency on nvrhi changes: https://github.com/RobertBeckebans/nvrhi/pull/6 for relaxing nvrhi push constant limits to permit platform-specific runtime checks.

Details are as follows:

1. Separates the single large constant buffer into renderparm subsets (12 in total: 3 of 128 bytes in size, 6 greater than 128 bytes but less than 256 bytes, and 3 greater than 256 bytes but less than 1024 bytes).
2. Adds new binding layout types to associate and differentiate between the new subsets (`BINDING_LAYOUT_GBUFFER, BINDING_LAYOUT_GBUFFER_SKINNED, BINDING_LAYOUT_TEXTURE, BINDING_LAYOUT_TEXTURE_SKINNED, BINDING_LAYOUT_WOBBLESKY, BINDING_LAYOUT_SSGI, BINDING_LAYOUT_SSGI_SKINNED, BINDING_LAYOUT_POST_PROCESS`).
3. Implements push constants for Vulkan and DX12 across all platforms: Linux, macOS, Windows. This has varying degrees of performance improvement, the largest being on Vulkan for Linux and macOS. Windows Vulkan shows modest improvement dependant on the GPU vendor (Nvidia's 256 byte limit is better than AMD's 128 byte limit on Windows). Windows DX12 shows similar performance when using push constants vs. volatile constant buffers. However, DX12 does show a performance improvement due to constant buffer subsetting with better change detection logic.  I have defined a new boolean `r_useDX12PushConstants` cvar which is turned off by default. This can optionally be turned on using autoexec.cfg for experimentation.
4. Reduced the volatile constant max buffer count from 16,384 to 8,192. I believe this is sufficient but if testing reveals differently, then it could be boosted back up. Note that when push constants are enabled it reduces the requirement.
5. Adds basic infrastructure for static constant buffers but these are disabled for now. This could be a possibility for the future but further subset refactoring would likely be needed, and sync issues would have to be resolved.
6. Modified uniforms change detection logic (orthogonal to push constants) which has a very significant positive impact on performance. See performance timings below. Also added new cvar `r_useVulkanPushConstants` (default on) which is useful for performance comparisons.
7. Don't allocate constant buffers unless required (i.e. when push constants disabled for a given binding layout type).
8. Fixed ImmediateMode (mainly for debug tools) to work with push constants enabled.

Tested on Windows 10 (AMD and Nvidia), Linux Manjaro, and macOS Ventura 13.5

Performance timings for this PR vs. current master, generated using a simple home-made timedemo:

**Windows Nvidia System (1070 Ti)**
DX12: 263 fps before, 360 fps after (with r_useDX12PushConstants = 0) --> significant improvement
Vulkan: 218 fps before, 333 fps after --> significant improvement

**Windows AMD System (6600 XT)**
DX12: 295 fps before, 305 fps after (with r_useDX12PushConstants = 0) --> neutral/positive improvement
Vulkan: 150 fps before, 160 fps after --> neutral/positive improvement

**Linux AMD System (6600 XT)**
Vulkan: 150 fps before, 270 fps after --> large improvement

**macOS AMD System (6600 XT)**
Vulkan: 77 fps before, 245 fps after --> very large improvement

**macOS Apple Silicon System (M1 Air)**
Vulkan: 6 fps before, 85 fps after --> massive improvement