@SRSaunders 2024-02-05
Now updated to your latest master with retro rendering modes and crt post processing.  The merge was large given the number of changes to master, but fairly straight forward with respect to the new shaders.  Fortunately they all fell into the `BINDING_LAYOUT_POST_PROCESS_FINAL` binding layout type with its existing renderparm subset.  I did not have to make any changes to the mappings, just simply mod the new shaders to add the `pc.` preface and include the correct subset.

Here is the updated mapping spreadsheet: [Binding to Shader Mapping v5.xlsx](https://github.com/RobertBeckebans/RBDOOM-3-BFG/files/14168499/Binding.to.Shader.Mapping.v5.xlsx)

At first I was not sure about the new modes, but now I really like the PSX + Newpixie CRT setting.  Feels very 90s!