@SRSaunders 2024-02-05
Thanks for the heads up re your PSX branch and a new renderparm.  Is there any way I could have a look at that to understand the implications for this branch and future subset handing?