@SRSaunders 2024-08-19
@RobertBeckebans I have updated this branch/PR to be compatible with your new PSX vertex jitter and affine rendering modes.  It was a bit more work than last time since the renderparms had changed for many shaders.  I had to do a bit of refactoring to make as much as possible fit into Vulkan push constant limits: 128 bytes for AMD on Windows, 256 bytes for nVidia on Windows/linux and AMD on linux.  macOS was not an issue since the push constant limit there is 4096 bytes.  I have attached the new renderparm to shader mapping spreadsheet here for anyone who cares to see how it works.

I am wondering what to do with this branch going forward.  I know you will likely not merge this into main, but I am putting it out there for other macOS, linux, and Windows users who want a bit more performance out of Vulkan.  Do you have the concept of optional components or branches (e.g. VR) that you may want to distribute in the future?  Perhaps this is something that could fit into that kind of optionality framework, at least for a fixed release version.

In any case, great work on the PSX rendermode stuff!  I find the PSX retro look combined with the New Pixie crt filter to be pretty cool.  Should be excellent as an engine for developing retro games with the original Tomb Raider look/style.