@SRSaunders 2024-08-20
> it is interesting what kind of an impact it has to just reference a few more renderparms besides the new rpPSXDistortions

This is mainly due to the heavy restriction on Vulkan push constant sizes in Windows and linux:  128 bytes for some implementations and 256 for others.  That's not a lot of data when you are doing innovative shader calculations that require many float parameters, matrices, etc.  It's a struggle to fit the parameters into these small limits.

> I would tend to keep your changes into a macOS branch.

Interesting thought.  Up to now I have been thinking about this problem as a general performance solution for Vulkan across all platforms.  If I were to look at a solution for macOS only, I might approach it differently since push constant limits are not a concern on that platform, and juggling to fit things into 128 or 256 bytes would not be necessary.  Size optimization is still useful there, but eliminating these low threshold limits would relax the need to worry about every single shader parm addition when doing new work.  In addition, if I were solving this for macOS only, the number of render parm sets could be reduced or eliminated and the general-purpose code for dynamic activation of push constant sets at runtime could be simplified.  If I were to come up with a new macOS-only design for this would you be interested?