@SRSaunders 2024-12-09
Updated [Binding to Shader Mapping v8.xlsx](https://github.com/user-attachments/files/18055999/Binding.to.Shader.Mapping.v8.xlsx) spreadsheet that includes the new Octahedron Cube binding layouts.