@klaussilveira 2025-02-10
With this PR applied, I no longer need to build the project on Debug in order to run in on Debian 12, RTX 4060 and latest Vulkan. This has also fixed issues when loading heavy GLTF scenes and models, as well as flying around in large levels with "noclip" on.