@RobertBeckebans 2025-03-04
Well we could introduce a console command that parses each shader and like the old HLSL -> GLSL converter that was removed from the renderer but in an easier fashion. It could build a list of used renderparms by checking "pc", "." and "renderparmX" and then pick which push constants subset is used. I don't want to use SPIRV-reflect as the code also affects the DX12 backend and it is possible to build the entire code base without Vulkan.