@SRSaunders 2025-05-12
Updated baseline to match 1.6.0 official release https://github.com/RobertBeckebans/RBDOOM-3-BFG/releases/tag/v1.6.0.

A big milestone! Congrats to @RobertBeckebans for all his hard work leading to significant capability improvements.

This PR is a set of changes on top of that which yields better Vulkan performance especially on macOS + Apple Silicon.  Linux users may also see some improvement depending on their machine + GPU.