@SRSaunders 2025-05-13
Thanks @RobertBeckebans, it's a fun project and I'm happy to contribute where I can.

Great news that you are considering this branch for your Flatpak build.  Please let me know if you need anything.