@SRSaunders 2025-05-16
@RobertBeckebans and @tomkidd:

I tried out the Universal Binary hosted by @tomkidd and noticed an issue when running on my x86 + AMD macOS machine.  I was erroneously performing a build-time check for host type vs. a runtime-check as part of a stability workaround when configuring push constants.  I had tested using single-architecture builds on x86_64 and arm64 and they both worked fine.  However, this did not work properly for macOS Universal Builds running on x86_64 hosts where the same source code has to work correctly for both architectures.  Testing the Universal Binary forced me to realize my mistake.

While the problem should only be visible for macOS Universal Builds running on x86_64 hosts, and not with single architecture builds on macOS and linux, you may want to pick up this revised version for your distributions.